auth/pyproject.toml
[tool.poetry]
name = "adv-auth"
version = "0.1.0"
description = ""
authors = ["Shoaib <iamshoaibxyz@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
email-validator = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-mail = "^1.4.1"
jinja2 = "^3.1.4"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.1.19"}
aiokafka = "^0.10.0"
protobuf = "^5.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
